Fix sign conversion warning in `Mine::miningEnabled`

+++ OPHD/Mine.cpp
@@ -89,7 +89,7 @@ std::bitset<4> Mine::miningEnabled() const
 
 void Mine::miningEnabled(OreType oreType, bool value)
 {
-	mFlags[static_cast<int>(oreType)] = value;
+	mFlags[static_cast<std::size_t>(oreType)] = value;
 }
